pyproject.toml
[project]
name = "fiveguys-cs490-project-summer-2025"
version = "0.1.0"
description = ""
authors = [
    {name = "Lior",email = "ultron895@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "firebase-admin (>=6.9.0,<7.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "flask-cors (>=6.0.0,<7.0.0)",
    "google-cloud-storage (>=3.1.0,<4.0.0)",
    "pymongo[srv] (>=4.13.0,<5.0.0)",
    "pdfplumber (>=0.11.7,<0.12.0)",
    "pdfminer-six (>=20250506,<20250507)",
    "docx2txt (>=0.9,<0.10)",
    "odfpy (>=1.4.1,<2.0.0)",
    "openai (>=1.86.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
